Add equipment selection step
